ai-market-intelligence-advanced/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
requests
textblob
cachetools
dnspython
